requirements.txt
Flask==2.2.3
Flask-Cors
Flask-APScheduler
cryptography==40.0.1
PyMySQL==1.0.3
pandas>=1.4.4
numpy>=1.21.5
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.13+cpu
torch_geometric>=2.3
tensorboard
scikit-learn>=1.1